planventure-api/requirements.txt
# API framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Database and ORM
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.40
flask-migrate==4.0.7

# Authentication and security
flask-jwt-extended==4.6.0
werkzeug==2.3.8
bcrypt==4.1.2

# Image processing
Pillow==10.4.0

# Object storage
minio==7.2.9

# Production server
gunicorn==21.2.0